Add work around to dome controller stopping before arriving in position.
We are currently investigating this issue, but in the meantime we will add this workaround.
Basically, for every move command the CSC starts a background task that monitors the position of the dome.
If the dome stops, the background task reissue the move command, and continues to monitor the position until the dome arrives.